refactored some functions to sample.py
